Rename xendit example identifier
