+++ app/src/main/java/com/xendit/example/AuthenticationActivity.java
@@ -1,4 +1,4 @@
-package ch.itomy.xendit_example;
+package com.xendit.example;
 
 import android.content.Context;
 import android.content.Intent;
@@ -15,10 +15,10 @@ import android.widget.Toast;
 
 import com.xendit.AuthenticationCallback;
 import com.xendit.Models.Authentication;
-import com.xendit.Models.Token;
 import com.xendit.Models.XenditError;
-import com.xendit.TokenCallback;
 import com.xendit.Xendit;
+
+import ch.itomy.xendit_example.R;
 
 /**
  * Created by Sergey on 4/3/17.
