+++ app/src/main/res/layout/activity_main.xml
@@ -5,7 +5,7 @@
     android:layout_width="match_parent"
     android:layout_height="match_parent"
     android:background="@android:color/white"
-    tools:context="ch.itomy.xendit_example.MainActivity">
+    tools:context="ch.itomy.xendit_example.com.xendit.example.MainActivity">
 
 
     <LinearLayout
