+++ app/src/test/java/com/xendit/example/ExampleUnitTest.java
@@ -1,4 +1,4 @@
-package ch.itomy.xendit_example;
+package com.xendit.example;
 
 import org.junit.Test;
